requirements.txt
streamlit
openai
dotenv
google-genai